client/requirements.txt
python-socketio[client]
Pillow
numpy
